Move GsonStrategies class to OAuth
