+++ dev/com.ibm.ws.security.common/bnd.bnd
@@ -70,8 +70,7 @@ instrument.classesExcludes: com/ibm/ws/security/common/internal/resources/*.clas
     com.ibm.ws.security.authentication;version=latest, \
     com.ibm.json4j;version=latest, \
     com.ibm.ws.webcontainer.security;version=latest, \
-    com.ibm.ws.org.apache.httpcomponents;version=latest, \
-    com.ibm.ws.kernel.boot.core;version=latest
+    com.ibm.ws.org.apache.httpcomponents;version=latest
 
 -testpath: \
     ../build.sharedResources/lib/junit/old/junit.jar;version=file, \
