+++ dev/com.ibm.ws.security.oauth/bnd.bnd
@@ -153,8 +153,8 @@ Include-Resource: \
 	com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
 	com.ibm.ws.security.openidconnect.clients.common;version=latest,\
 	com.ibm.ws.org.eclipse.equinox.metatype;version=latest,\
-	com.ibm.ws.security.jwt;version=latest
-
+	com.ibm.ws.security.jwt;version=latest,\
+	com.ibm.ws.kernel.boot.core;version=latest
 
 -testpath: \
     ../build.sharedResources/lib/junit/old/junit.jar;version=file, \
